requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
aiomysql==0.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
alembic==1.7.1
pytest==6.2.5
pytest-asyncio==0.15.1
selenium==3.141.0
locust==2.8.3
python-dotenv==0.19.0
pydantic==1.8.2
email-validator==1.1.3
aiofiles==0.7.0
python-slugify==5.0.2
pytest-cov==2.12.1 